+++ src/frontend/apps/desk/src/features/mail-domains/components/index.tsx
@@ -1,0 +1,2 @@
+export * from './MailDomainsContent';
+export * from './MailDomainsLayout';
