--- src/frontend/apps/desk/src/features/mails/components/index.tsx
@@ -1,2 +1,0 @@
-export * from './MailContent';
-export * from './MailLayout';
