+++ src/frontend/apps/e2e/__tests__/app-desk/mail-domains.spec.ts
